Move state2vec and vec2state to a utility file, removed copy-pasted code
